pyproject.toml
[project]
name = "mcp-discord"
version = "0.1.0"
description = "MCP Server for Discord"
authors = [
    {name = "Hanweg Altimer", email = "hanwegaltimer@outlook.com"},
    {name = "Dylan Huang", email = "dhuang@fireworks.ai"}
]
dependencies = [
    "discord.py>=2.3.0",
    "fastmcp>=2.10.6"
]
requires-python = ">=3.10"
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
Homepage = "https://github.com/eval-protocol/mcp-discord"
Repository = "https://github.com/eval-protocol/mcp-discord.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["mcp_discord"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
mcp-discord = "mcp_discord:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
